+++ src/main/java/me/melontini/recipebookispain/mixin/MouseWheelieCompatMixin.java
@@ -2,15 +2,11 @@ package me.melontini.recipebookispain.mixin;
 
 import de.siphalor.mousewheelie.client.mixin.gui.other.MixinRecipeBookWidget;
 import de.siphalor.mousewheelie.client.util.ScrollAction;
-import me.melontini.recipebookispain.access.RecipeBookWidgetAccess;
-import me.melontini.recipebookispain.access.RecipeGroupButtonAccess;
+import net.minecraft.client.gui.screen.recipebook.RecipeBookResults;
 import net.minecraft.client.gui.screen.recipebook.RecipeBookWidget;
 import net.minecraft.client.gui.screen.recipebook.RecipeGroupButtonWidget;
 import org.jetbrains.annotations.Nullable;
-import org.spongepowered.asm.mixin.Dynamic;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Pseudo;
-import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.*;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -23,17 +19,26 @@ public abstract class MouseWheelieCompatMixin {
     @Nullable
     private RecipeGroupButtonWidget currentTab;
 
+    @Shadow @Final private RecipeBookResults recipesArea;
+
     @SuppressWarnings("ReferenceToMixin")
     @Dynamic(mixin = MixinRecipeBookWidget.class)
-    @Inject(at = @At(value = "INVOKE", target = "net/minecraft/client/gui/screen/recipebook/RecipeGroupButtonWidget.setToggled (Z)V", ordinal = 1, shift = At.Shift.AFTER), method = "mouseWheelie_scrollRecipeBook")
+    @Inject(at = @At(value = "INVOKE", target = "net/minecraft/client/gui/screen/recipebook/RecipeGroupButtonWidget.setToggled (Z)V", ordinal = 1, shift = At.Shift.AFTER), method = "mouseWheelie_scrollRecipeBook", require = 0)
     private void rbip$scrollPages(double mouseX, double mouseY, double scrollAmount, CallbackInfoReturnable<ScrollAction> cir) {
-        RecipeBookWidget bookWidget = (RecipeBookWidget) (Object) this;
-        RecipeBookWidgetAccess widgetAccess = ((RecipeBookWidgetAccess) bookWidget);
-        RecipeGroupButtonAccess buttonAccess = ((RecipeGroupButtonAccess) this.currentTab);
+        if (this.currentTab == null) return;//how tho?
+        this.recipesArea.hideAlternates();
 
-        assert buttonAccess != null;
-        if (widgetAccess.getPage() != buttonAccess.getPage()) {
-            widgetAccess.setPage(buttonAccess.getPage());
+        RecipeBookWidget bookWidget = (RecipeBookWidget) (Object) this;
+        if (bookWidget.dm$getPage() != currentTab.dm$getPage()) {
+            bookWidget.dm$setPage(currentTab.dm$getPage());
         }
     }
+
+    @SuppressWarnings("ReferenceToMixin")
+    @Dynamic(mixin = MixinRecipeBookWidget.class)
+    @Inject(at = @At(value = "INVOKE", target = "de/siphalor/mousewheelie/client/util/accessors/IRecipeBookResults.mouseWheelie_refreshResultButtons ()V", shift = At.Shift.BEFORE), method = "mouseWheelie_scrollRecipeBook", require = 0)
+    private void rbip$hideAlternatives(double mouseX, double mouseY, double scrollAmount, CallbackInfoReturnable<ScrollAction> cir) {
+        this.recipesArea.hideAlternates();
+    }
+
 }